+++ src/components/filter/dualSlider/dual-slider.scss
